Add support for XinJE protocol.
There does not seem to be a formal description of this protocol, so this
current implementation has relied on packet sniffing to determine the
actual use of the protocol.

The XinJE protocol is accessed via the Modbus function codes 30-33, and
so can be considered to be an extension of Modbus. It functions very
similarly to Modbus, but with a few key differences:

- Addresses are 32 bits rather than 16.

- Coils are not read as individual bits of a larger 8 byte word, but
  instead are read one at a time, always in the LSB position of the
  yielded word. This means that addressing a coil is simple, as the
  coil's address can directly be used to query or set its bit.

- There are not separate function codes for ro and r/w coils/registers.

Function codes 30 and 31 respectively read from and write to coils,
while function codes 32 and 33 handle registers. Therefore, the value of
30 is used here for both reading and writing with regard to coils, with
32 being used for registers.